Implement mov biasing (#10)
* Change get-color method in dsatur

AND WIP: mov biasing

* Disable workflows

* Fuzzing changes

* Fuzzing continues

* Use sets when getting saturation

The saturation lists might have the same colour multiple times. So
put it in a set so the the get-color algorithm can be successful.

Previously it was assigning wrong colours, when sat list was
'(1 2 3 3 4) for example, it would assign the colour 3 when though
it is conflicting. This fixes it.

* Make get colour more resilient

* Slurm mail on Job end, not begin

* Implement mov biasing

* Don't use the reserved registers as mov related